ScaledImageFactory: Move mJobPool to stack implementation
Using a stack means the workers render the most recent requests
first, which leads to a more responsive UI if the user panning gets
ahead of the worker threads. There's still a case where you can pan to
a new area, then pan back to the old one, and the job-queue isn't smart
enough to deal with that yet, but this covers keeping the UI smooth while
scrolling only to new area of the image.

Also, implements wxMsgStack.h, heavily based on wx/msgqueue.h